requirements.txt
Adafruit-Blinka==8.0.1
adafruit-circuitpython-busdevice==5.1.9
adafruit-circuitpython-neopixel==6.3.1
adafruit-circuitpython-pixelbuf==1.1.4
adafruit-circuitpython-typing==1.7.1
Adafruit-PlatformDetect==3.24.1
Adafruit-PureIO==1.1.9
asttokens==2.0.5
backcall==0.2.0
blessings==1.7
board==1.0
bpython==0.22.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorzero==2.0
curtsies==0.3.10
cwcwidth==0.1.6
decorator==5.1.1
Encoder==1.1
executing==0.8.3
gpiozero==1.6.2
greenlet==1.1.2
grove.py==0.6
i2cdevice==0.0.7
idna==3.3
ipython==8.4.0
jedi==0.18.1
lsm303d==0.0.5
matplotlib-inline==0.1.3
numpy==1.22.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pigpio==1.78
Pillow==9.1.1
pip==22.1.2
pkg-resources==0.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pyftdi==0.54.0
Pygments==2.12.0
pynmea2==1.18.0
pyserial==3.5
pyusb==1.2.1
pyxdg==0.28
requests==2.27.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
setuptools==62.3.2
six==1.16.0
smbus2==0.4.2
smbus==1.1.post2
spidev==3.5
stack-data==0.2.0
sysv-ipc==1.1.0
traitlets==5.2.2.post1
typing-extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
